Day 10 complete
I didn’t use Event.shiftKey (because I didn’t know it was a thing) and
I swapped out the divs for labels. However, shift + clicking on a label
doesn’t seem to work the same as shift + clicking on a checkbox
directly. Even if I totally remove the JS from this file. Odd.