Day 12 update, secrete is now an array
Arguably easier to deal with keys with multi-character values this wahy.

Interestingly, the cornify.js script has a check for the Konami code
baked in. And it uses keycodes. *But* it concats those into a string.

That section of code is in turn based on another project, hosted on
GitHub here: https://github.com/snaptortoise/konami-js